+++ src/ByteSizeLib.Tests/ByteSizeLib.Tests.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="../ByteSize/ByteSize.csproj" />
+    <ProjectReference Include="../ByteSizeLib/ByteSizeLib.csproj" />
   </ItemGroup>
 
   <ItemGroup>
